Move OpenGL Material class to a shared name-space
OpenGL Materials are actually shared between Windows and Linux
implementation and there is no need to be duplicated
Also extracted common materials into static fields of `OpenGLMaterial`.
The rendering abstraction will be further refactored and this is the
first step.
