+++ src/ImGui/OSImplentation/Windows/Win32OpenGLRenderer.cs
@@ -4,157 +4,15 @@ using ImGui.Rendering;
 using System;
 using System.Collections.Generic;
 using System.Runtime.InteropServices;
+using ImGui.OSImplentation.Shared;
 
 namespace ImGui.OSImplentation.Windows
 {
     internal partial class Win32OpenGLRenderer : IRenderer
     {
-        public readonly OpenGLMaterial shapeMaterial = new OpenGLMaterial(
-            vertexShader: @"
-#version 330
-uniform mat4 ProjMtx;
-in vec2 Position;
-in vec2 UV;
-in vec4 Color;
-out vec2 Frag_UV;
-out vec4 Frag_Color;
-void main()
-{
-	Frag_UV = UV;
-	Frag_Color = Color;
-	gl_Position = ProjMtx * vec4(Position.xy,0,1);
-}
-",
-            fragmentShader: @"
-#version 330
-uniform sampler2D Texture;
-in vec2 Frag_UV;
-in vec4 Frag_Color;
-out vec4 Out_Color;
-void main()
-{
-	Out_Color = Frag_Color;
-}
-"
-            );
-
-        public readonly OpenGLMaterial imageMaterial = new OpenGLMaterial(
-            vertexShader: @"
-#version 330
-uniform mat4 ProjMtx;
-in vec2 Position;
-in vec2 UV;
-in vec4 Color;
-out vec2 Frag_UV;
-out vec4 Frag_Color;
-void main()
-{
-	Frag_UV = UV;
-	Frag_Color = Color;
-	gl_Position = ProjMtx * vec4(Position.xy,0,1);
-}
-",
-            fragmentShader: @"
-#version 330
-uniform sampler2D Texture;
-in vec2 Frag_UV;
-in vec4 Frag_Color;
-out vec4 Out_Color;
-void main()
-{
-	Out_Color = Frag_Color * texture( Texture, Frag_UV.st);
-}
-"
-            );
-
-        public static readonly OpenGLMaterial glyphMaterial = new OpenGLMaterial(
-    vertexShader: @"
-#version 330
-uniform mat4 ProjMtx;
-in vec2 Position;
-in vec2 UV;
-in vec4 Color;
-out vec2 Frag_UV;
-out vec4 Frag_Color;
-uniform vec2 offset;
-void main()
-{
-	Frag_UV = UV;
-	Frag_Color = Color;
-	gl_Position = ProjMtx * vec4(offset+Position.xy,0,1);
-}
-",
-    fragmentShader: @"
-#version 330
-in vec2 Frag_UV;
-in vec4 Frag_Color;
-out vec4 Out_Color;
-uniform vec4 color;
-void main()
-{
-	if (Frag_UV.s * Frag_UV.s - Frag_UV.t > 0.0)
-	{
-		discard;
-	}
-
-	// Upper 4 bits: front faces
-	// Lower 4 bits: back faces
-	Out_Color = Frag_Color* color *0.001 + vec4(0.125, 0, 0, 0); /*Frag_Color* color * (gl_FrontFacing ? 16.0 / 255.0 : 1.0 / 255.0);*/
-    Out_Color.a = 1;
-}
-"
-    );
-        public static readonly OpenGLMaterial textMaterial = new OpenGLMaterial(
-    vertexShader: @"
-#version 330
-in vec2 Position;
-in vec2 UV;
-in vec4 Color;
-out vec2 Frag_UV;
-out vec4 Frag_Color;
-void main()
-{
-	Frag_UV = UV;
-    Frag_Color = Color;
-	gl_Position = vec4(Position, 0.0, 1.0);
-}
-",
-    fragmentShader: @"
-#version 330
-in vec2 Frag_UV;
-in vec4 Frag_Color;
-out vec4 Out_Color;
-uniform sampler2D Texture;
-void main()
-{
-	// Get samples for -2/3 and -1/3
-	vec2 valueL = texture2D(Texture, vec2(Frag_UV.x + dFdx(Frag_UV.x), Frag_UV.y)).yz * 255.0;
-	vec2 lowerL = mod(valueL, 16.0);
-	vec2 upperL = (valueL - lowerL) / 16.0;
-	vec2 alphaL = min(abs(upperL - lowerL), 2.0);
-
-	// Get samples for 0, +1/3, and +2/3
-	vec3 valueR = texture2D(Texture, Frag_UV).xyz * 255.0;
-	vec3 lowerR = mod(valueR, 16.0);
-	vec3 upperR = (valueR - lowerR) / 16.0;
-	vec3 alphaR = min(abs(upperR - lowerR), 2.0);
-
-	// Average the energy over the pixels on either side
-	vec4 rgba = vec4(
-		(alphaR.x + alphaR.y + alphaR.z) / 6.0,
-		(alphaL.y + alphaR.x + alphaR.y) / 6.0,
-		(alphaL.x + alphaL.y + alphaR.x) / 6.0,
-		0.0);
-
-	Out_Color = vec4(rgba.rgb,1) + Frag_Color*0.001;// + vec4(0,1,0,1);
-}
-"
-    );
-
         //Helper for some GL functions
         private static readonly int[] IntBuffer = { 0, 0, 0, 0 };
         private static readonly float[] FloatBuffer = { 0, 0, 0, 0 };
-        private static readonly uint[] UIntBuffer = { 0, 0, 0, 0 };
 
         //framebuffer
         private readonly uint[] framebuffers = { 0 };
@@ -163,7 +21,6 @@ void main()
         static uint framebuffer = 0;
         static uint framebufferColorTexture = 0;
         static QuadMesh quadMesh;
-        static ITexture checkerTexture;
 
         static readonly Point[] JITTER_PATTERN =
         {
@@ -177,11 +34,7 @@ void main()
         public void Init(IntPtr windowHandle, Size size)
         {
             CreateOpenGLContext(windowHandle);
-
-            this.shapeMaterial.Init();
-            this.imageMaterial.Init();
-            glyphMaterial.Init();
-            textMaterial.Init();
+            OpenGLMaterial.InitCommonMaterials();
 
             // Other state
             GL.Enable(GL.GL_MULTISAMPLE);
@@ -226,8 +79,6 @@ void main()
             GL.BindFramebuffer(GL.GL_FRAMEBUFFER_EXT, 0);
 
             quadMesh = new QuadMesh(size.Width, size.Height);
-            //checkerTexture = TextureCache.Default.GetOrAdd(
-            //    Environment.GetFolderPath(Environment.SpecialFolder.Desktop) + "/checker.jpg");
             Utility.CheckGLError();
         }
 
@@ -239,8 +90,8 @@ void main()
 
         public void DrawMeshes(int width, int height, (Mesh shapeMesh, Mesh imageMesh, TextMesh textMesh) meshes)
         {
-            DrawMesh(this.shapeMaterial, meshes.shapeMesh, width, height);
-            DrawMesh(this.imageMaterial, meshes.imageMesh, width, height);
+            DrawMesh(OpenGLMaterial.shapeMaterial, meshes.shapeMesh, width, height);
+            DrawMesh(OpenGLMaterial.imageMaterial, meshes.imageMesh, width, height);
             DrawTextMesh(meshes.textMesh, width, height);
         }
 
@@ -434,6 +285,7 @@ void main()
             {
                 return;
             }
+            var glyphMaterial = OpenGLMaterial.glyphMaterial;
 
             // Backup GL state
             GL.GetIntegerv(GL.GL_CURRENT_PROGRAM, IntBuffer); int last_program = IntBuffer[0];
@@ -551,9 +403,7 @@ void main()
 
         public void ShutDown()
         {
-            this.shapeMaterial.ShutDown();
-            this.imageMaterial.ShutDown();
-            glyphMaterial.ShutDown();
+            OpenGLMaterial.ShutDownCommonMaterials();
         }
 
         public byte[] GetRawBackBuffer(out int width, out int height)
